// frontend/src/app/models/a9522b_ext_imei.model.ts
export class a9522b_ext_imei {
    id9522b_ext_imei: number;
    imei: string;
    sn: string;
    remodel: string;
}

export interface a9522b_ext_imeiAll {
    result: a9522b_ext_imei[],
    message: string
}

export interface a9522b_ext_imeiBy{
    result: a9522b_ext_imei,
    message: string
}