// frontend/src/app/models/a9523_ext_imei.model.ts
export class a9523_ext_imei {
    id9523_ext_imei: number;
    imei: string;
    sn: string;
    remodel: string;
}

export interface a9523_ext_imeiAll {
    result: a9523_ext_imei[],
    message: string
}

export interface a9523_ext_imeiBy{
    result: a9523_ext_imei,
    message: string
}