// frontend/src/app/models/a9523mw_imei.model.ts
export class a9523mw_imei {
    id9523mw_imei: number;
    imei: string;
    sn: string;
    remodel: string;
}

export interface a9523mw_imeiAll {
    result: a9523mw_imei[],
    message: string
}

export interface a9523mw_imeiBy{
    result: a9523mw_imei,
    message: string
}