// frontend/src/app/models/a9523ptt_imei.model.ts
export class a9523ptt_imei {
    id9523ptt_imei: number;
    imei: string;
    sn: string;
    remodel: string;
}

export interface a9523ptt_imeiAll {
    result: a9523ptt_imei[],
    message: string
}

export interface a9523ptt_imeiBy{
    result: a9523ptt_imei,
    message: string
}