// frontend/src/app/models/a9555a_ext_imei.model.ts
export class a9555a_ext_imei {
    id9555a_ext_imei: number;
    imei: string;
    sn: string;
    remodel: string;
}

export interface a9555a_ext_imeiAll {
    result: a9555a_ext_imei[],
    message: string
}

export interface a9555a_ext_imeiBy{
    result: a9555a_ext_imei,
    message: string
}