// frontend/src/app/models/a9555a_gsa_imei.model.ts
export class a9555a_gsa_imei {
    id9555a_gsa_imei: number;
    imei: string;
    sn: string;
    remodel: string;
}

export interface a9555a_gsa_imeiAll {
    result: a9555a_gsa_imei[],
    message: string
}

export interface a9555a_gsa_imeiBy{
    result: a9555a_gsa_imei,
    message: string
}