// frontend/src/app/models/a9575_ext_imei.model.ts
export class a9575_ext_imei {
    id9575_ext_imei: number;
    imei: string;
    sn: string;
    remodel: string;
}

export interface a9575_ext_imeiAll {
    result: a9575_ext_imei[],
    message: string
}

export interface a9575_ext_imeiBy{
    result: a9575_ext_imei,
    message: string
}