// frontend/src/app/models/a9575gsa_imei.model.ts
export class a9575gsa_imei {
    id9575gsa_imei: number;
    imei: string;
    sn: string;
    remodel: string;
}

export interface a9575gsa_imeiAll {
    result: a9575gsa_imei[],
    message: string
}

export interface a9575gsa_imeiBy{
    result: a9575gsa_imei,
    message: string
}