// frontend/src/app/models/a9575ptt_ext_imei.model.ts
export class a9575ptt_ext_imei {
    id9575ptt_ext_imei: number;
    imei: string;
    sn: string;
    remodel: string;
}

export interface a9575ptt_ext_imeiAll {
    result: a9575ptt_ext_imei[],
    message: string
}

export interface a9575ptt_ext_imeiBy{
    result: a9575ptt_ext_imei,
    message: string
}